backend/requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
numpy
pypdf
langchain-text-splitters
fastembed
groq
python-dotenv
python-docx